uucore: dedup code of unbounded splice with broker
